# slight optimization for the USB transport