--- a/pyxcp/transport/eth.py
+++ b/pyxcp/transport/eth.py
@@ -185,7 +185,6 @@ class Eth(BaseTransport):
                             break
                     else:
                         if current_size >= length:
-                            # response = memoryview(data[current_position : current_position + length])
                             response = data[current_position : current_position + length]
                             processResponse(response, length, counter, timestamp)