## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [x] I have read the **CONTRIBUTING** document.
- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.

```python
from array import array
buffer = array("B", bytes(16384))
buffer_view = memoryview(buffer)
%timeit buffer.tobytes()[:1024]
348 ns ± 10.4 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
%timeit buffer_view[:1024].tobytes()
177 ns ± 3.23 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
%timeit buffer[:1024].tobytes()
229 ns ± 6.94 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```